smart_attend/smartattend_backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
Pillow==11.3.0
django-cors-headers==4.9.0